% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hergm.R
\name{arrange_edgelist}
\alias{arrange_edgelist}
\title{Make the given edgelist consistent with the data frame that contains vertex info.}
\usage{
arrange_edgelist(edgelist, sorted_dataframe)
}
\arguments{
\item{edgelist}{an edgelist to be arranged}

\item{sorted_dataframe}{a data frame sorted by \code{sort_block_membership}}
}
\description{
Make the given edgelist consistent with the data frame that contains vertex info.
}
