% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_within_params.R
\name{as_sparse_adj}
\alias{as_sparse_adj}
\title{Get a sparse adjacency matrix from a network object}
\usage{
as_sparse_adj(net)
}
\arguments{
\item{net}{a network object}
}
\description{
Get a sparse adjacency matrix from a network object
}
