% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hergm.R
\name{combine_within_between_edges}
\alias{combine_within_between_edges}
\title{Combine within- and between-block edges while removing duplicated links.}
\usage{
combine_within_between_edges(
  edgelist_within,
  edgelist_between,
  use_fast_between_simulation
)
}
\arguments{
\item{edgelist_within}{an within-block edgelist}

\item{edgelist_between}{a between-block edgelist (Potentially, there are edges that also appear in the within-block edgelist)}

\item{use_fast_between_simulation}{If \code{TRUE}, this function uses an effcient way to simulate a between-block network. If the network is very large, you should consider using this option.}
}
\description{
Combine within- and between-block edges while removing duplicated links.
}
