% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_sparse_feature_adjmat.R
\name{compute_multiplied_feature_matrices}
\alias{compute_multiplied_feature_matrices}
\title{Get a list of sparse feature adjacency matrix from a formula}
\usage{
compute_multiplied_feature_matrices(net, list_feature_matrices)
}
\arguments{
\item{net}{a network object from which nodal covariates are extracted.}

\item{list_feature_matrices}{a list of feature adjacency matrices generated by \code{get_list_sparse_feature_adjmat()}.}
}
\description{
Get a list of sparse feature adjacency matrix from a formula
}
