% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_yule_coef}
\alias{compute_yule_coef}
\title{Compute Yule's Φ-coefficient}
\usage{
compute_yule_coef(z_star, z)
}
\arguments{
\item{z_star}{a true block membership}

\item{z}{an estimated block membership}
}
\description{
Compute Yule's Φ-coefficient
}
