% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hergm.R
\name{edgelist_to_stats}
\alias{edgelist_to_stats}
\title{Converts an edgelist into a matrix of sufficient network statistics}
\usage{
edgelist_to_stats(net, edgelist, between_formula)
}
\arguments{
\item{net}{the net to extract the covariates from}

\item{edgelist}{the edgelist}

\item{between_formula}{the formula for the between connections}
}
\value{
a matrix of sufficient network statistics
}
\description{
Converts an edgelist into a matrix of sufficient network statistics
}
