% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_between_param.R
\name{estimate_between_param}
\alias{estimate_between_param}
\title{Estimate between-block parameters by logit}
\usage{
estimate_between_param(formula, network, block)
}
\arguments{
\item{formula}{formula for estimating between-block parameters}

\item{network}{network object}

\item{block}{a vector that represents which node belongs to which node}
}
\description{
Estimate between-block parameters by logit
}
