% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hergm.R
\name{extract_covariate_names}
\alias{extract_covariate_names}
\title{Extract Covariate Names
Extracts the names of covariates used in the formula}
\usage{
extract_covariate_names(formula_for_simulation)
}
\arguments{
\item{formula_for_simulation}{the formula to check}
}
\value{
A list of covariate names (can be empty)
}
\description{
Extract Covariate Names
Extracts the names of covariates used in the formula
}
