% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hergm.R
\name{generate_network_for_output}
\alias{generate_network_for_output}
\title{Create a final network}
\usage{
generate_network_for_output(
  formula_for_simulation_within,
  formula_for_simulation_between,
  sorted_dataframe,
  edgelist
)
}
\arguments{
\item{formula_for_simulation_within}{formula for simulating a within network}

\item{formula_for_simulation_between}{formula for simulating a between network}

\item{sorted_dataframe}{a data frame generated by \code{sort_block_membership}}

\item{edgelist}{an edgelist that contain both within- and between-block edges without duplication}
}
\description{
Create a final network
}
