% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hergm.R
\name{generate_seed_network}
\alias{generate_seed_network}
\title{Create a seed network from which a network will be simulated.}
\usage{
generate_seed_network(
  formula_for_simulation,
  sorted_dataframe,
  edgelist = NULL,
  directed
)
}
\arguments{
\item{formula_for_simulation}{formula for simulating a network}

\item{sorted_dataframe}{a data frame generated by \code{sort_block_membership}}

\item{edgelist}{an edgelist from which a seed network is generated. The class of the edgelist should be "edgelist", i.e. it should contain as attributes the number of nodes and vertex names.}

\item{directed}{a boolean of whether the network is directed.}
}
\description{
Create a seed network from which a network will be simulated.
}
