% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hergm.R
\name{get_between_stats}
\alias{get_between_stats}
\title{Converts a list of edgelists into a data frame of network statistics}
\usage{
get_between_stats(edgelists, between_formula)
}
\arguments{
\item{edgelists}{the list of edgelists}

\item{between_formula}{the formula for the between connections}
}
\value{
a data frame of sufficient network statistics
}
\description{
Converts a list of edgelists into a data frame of network statistics
}
