% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_lighthergm.R
\name{get_gof_stats}
\alias{get_gof_stats}
\title{Gets the GOF stats for a formula
If a network is passed, that one is used to obtain the network statistics,
otherwise the netwok in the formula is used.}
\usage{
get_gof_stats(
  sim_formula,
  net = NULL,
  sim_number = NULL,
  compute_geodesic_distance = FALSE
)
}
\arguments{
\item{sim_formula}{a formula}

\item{net}{a statnet network object}

\item{sim_number}{the ID of the current simulation}

\item{compute_geodesic_distance}{if TRUE, includes the geodesic distance in the result object}
}
\value{
a list with the goodness-of-fit statistics
}
\description{
Gets the GOF stats for a formula
If a network is passed, that one is used to obtain the network statistics,
otherwise the netwok in the formula is used.
}
