% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_lighthergm.R
\name{gof_lighthergm_within}
\alias{gof_lighthergm_within}
\title{Goodness of fit statistics for HERGM}
\usage{
gof_lighthergm_within(
  net,
  data_for_simulation,
  colname_vertex_id,
  colname_block_membership,
  lighthergm_results,
  ergm_control = ergm::control.simulate.formula(),
  seed = NULL,
  n_sim = 1,
  compute_geodesic_distance = FALSE,
  ...
)
}
\arguments{
\item{net}{the target network}

\item{data_for_simulation}{a dataframe with node-level covariates}

\item{colname_vertex_id}{the name of the column that contains the node id}

\item{colname_block_membership}{the name o the column that contains the block affiliation of each node}

\item{lighthergm_results}{a lighthergm results object}

\item{ergm_control}{MCMC parameters as an instance of ergm.control}

\item{seed}{the seed to be passed to simulate_hergm}

\item{n_sim}{the number of simulations to employ for calculating goodness of fit}

\item{compute_geodesic_distance}{if \code{TRUE}, the distribution of geodesic distances is also computed (considerably increases computation time on large networks. \code{FALSE} by default.)}

\item{...}{Additional arguments, to be passed to lower-level functions}
}
\description{
Goodness of fit statistics for HERGM
}
