% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_wrapper.R
\name{hergm.getnetwork}
\alias{hergm.getnetwork}
\title{a function that gets a network object from a formula}
\usage{
hergm.getnetwork(form, n_clusters, loopswarning = TRUE)
}
\arguments{
\item{form}{a formula that contains a network object in the left-hand side}

\item{n_clusters}{the maximum number of blocks}

\item{loopswarning}{if TRUE, this function warns that the network contains loops.}
}
\description{
a function that gets a network object from a formula
}
