% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_lighthergm.R
\name{separate_formulas}
\alias{separate_formulas}
\title{Separates a formula into its between and within components. The between component excludes
terms which introduce dyadic dependence.}
\usage{
separate_formulas(target_formula)
}
\arguments{
\item{target_formula}{a target formula}
}
\value{
a list containing the between and within formulas
}
\description{
Separates a formula into its between and within components. The between component excludes
terms which introduce dyadic dependence.
}
