% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hergm.R
\name{sort_block_membership}
\alias{sort_block_membership}
\title{Create a data frame from block memberships, vertex ids, and vertex covariates, sorted by block ids.}
\usage{
sort_block_membership(
  data_for_simulation,
  colname_vertex_id,
  colname_block_membership
)
}
\arguments{
\item{data_for_simulation}{a data frame that contains vertex id, block membership, and vertex features.}

\item{colname_vertex_id}{a column name in the data frame for the vertex id}

\item{colname_block_membership}{a column name in the data frame for the block membership.}
}
\description{
Create a data frame from block memberships, vertex ids, and vertex covariates, sorted by block ids.
}
