% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_lighthergm.R
\name{to_edgewise_shared_partners_df}
\alias{to_edgewise_shared_partners_df}
\title{Extracts the edgewise shared partners distribution (undirected).}
\usage{
to_edgewise_shared_partners_df(net)
}
\arguments{
\item{net}{a statnet network object}
}
\value{
a data frame
}
\description{
Extracts the edgewise shared partners distribution (undirected).
}
