% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_lighthergm.R
\name{to_geodesic_dist_df}
\alias{to_geodesic_dist_df}
\title{Extracts the geodesic distance distribution from a network and returns it as a dataframe.}
\usage{
to_geodesic_dist_df(net)
}
\arguments{
\item{net}{a statnet network object}
}
\value{
a data frame
}
\description{
Extracts the geodesic distance distribution from a network and returns it as a dataframe.
}
