% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyNet}
\alias{toyNet}
\title{A toy network to play \code{lighthergm} with.}
\format{
A \code{statnet}'s network class object. It has three nodal features.
\describe{
\item{block}{block membership of each node}
\item{x}{a covariate. It has 10 labels.}
\item{y}{a covariate. It has 10 labels.}
...
}
\code{x} and \code{y} are not variables with any particular meaning.
}
\usage{
toyNet
}
\description{
This network has a clear cluster structure.
The number of clusters is four, and which cluster each node belongs to is defined in the variable "block".
}
\keyword{datasets}
