% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_wrapper.R
\name{spec_clust_sparse}
\alias{spec_clust_sparse}
\title{function for spectral clustering}
\usage{
spec_clust_sparse(network, n_clusters, eigenvectors_fn)
}
\arguments{
\item{network}{a sparse adjacency matrix}

\item{n_clusters}{number of specified clusters}

\item{eigenvectors_fn}{a function that performs eigenvector decomposition}
}
\description{
function for spectral clustering
}
